+++ actionpack/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Set the default `ActionView::Digestor.cache` to `ActiveSupport::Cache::NullStore.new`
+    if `ActionView::Base.cache_template_loading` is disabled.
+
+    *Sean Huber*
+
 *   Fix an issue where partials with a number in the filename weren't being digested for cache dependencies.
 
     *Bryan Ricker*